Moved LoadingCircle to components folder
